Move parser package to io & rename
Was not under io package previously

Parsing of incoming inputs should be under io package
